# Speed up selector generation