--- a/src/web/element.ts
+++ b/src/web/element.ts
@@ -99,29 +99,3 @@ export const getTopmostEditableElement = (
   // This should never be hit, but here as a safety
   return element;
 };
-
-export const getElementText = (element: HTMLElement): string | null => {
-  let text = (element.innerText || '').trim();
-
-  if (
-    element instanceof HTMLInputElement &&
-    ['button', 'submit'].includes(element.type)
-  ) {
-    text = element.value;
-  }
-
-  if (
-    !text ||
-    text.length > 200 ||
-    text.match(/[\n\r\t]+/) ||
-    // ignore invisible characters which look like an empty string but have a length
-    // https://www.w3resource.com/javascript-exercises/javascript-string-exercise-32.php
-    // https://stackoverflow.com/a/21797208/230462
-    text.match(/[^\x20-\x7E]/g)
-  )
-    return null;
-
-  console.debug(`qawolf: found text="${text}" for element`, element);
-
-  return text;
-};