--- a/src/web/types.ts
+++ b/src/web/types.ts
@@ -9,6 +9,8 @@ export type ParsedSelector = {
 };
 
 export type Evaluator = {
+  createTextSelector(element: Element): string | undefined;
+
   isVisible(element: Element): boolean;
 
   querySelectorAll(selector: ParsedSelector, root: Node): HTMLElement[];