Resolves #776

Computational complexity:

buildCueSets: textSelectorsInTree! but there should not be too many text selectors in the dom tree

findBestCueGroup: ![image](https://user-images.githubusercontent.com/1136652/90097010-2fc72c80-dcf2-11ea-9bf0-e00e206256d3.png). Takes about 20ms on my machine for 700 combinations (n~10 & k=5).

o = buildCueSets * findBestCueGroup

On average: 2 sets * 700 combos (20ms) = 40ms
Hard limit: 50 sets * 7000 combos (100ms) = 5s